fixed#116 fixed issue while saving empty Quick fill

+++ menuplanner/grails-app/controllers/com/mp/domain/QuickFillController.groovy
@@ -21,18 +21,23 @@ class QuickFillController {
     }
 
     def saveAndUpdateQuickFill = {
+        if(!params?.quickFillName){
+            flash.message = "Please specify the name for the Quick Fill"
+            redirect(action: "quickFillAdmin")
+            return
+        }
         QuickFill quickFill = (params.id) ? QuickFill.get(params.long("id")) : new QuickFill()
         if (params.id) {
             def mealItems = quickFill.mealItems
             quickFill.mealItems = []
             mealItems*.delete()
         }
-        Meal meal1 = new Meal(type: MealType.DINNER, items: Item.getAll([params.meal.Dinner1].flatten()))
-        Meal meal2 = new Meal(type: MealType.DINNER, items: Item.getAll([params.meal.Dinner2].flatten()))
-        Meal meal3 = new Meal(type: MealType.DINNER, items: Item.getAll([params.meal.Dinner3].flatten()))
-        Meal meal4 = new Meal(type: MealType.DINNER, items: Item.getAll([params.meal.Dinner4].flatten()))
-        Meal meal5 = new Meal(type: MealType.BREAKFAST, items: Item.getAll([params.meal.Breakfast].flatten()))
-        Meal meal6 = new Meal(type: MealType.LUNCH, items: Item.getAll([params.meal.Lunch].flatten()))
+        Meal meal1 = new Meal(type: MealType.DINNER, items: Item.getAll([params?.meal?.Dinner1]?.flatten()))
+        Meal meal2 = new Meal(type: MealType.DINNER, items: Item.getAll([params?.meal?.Dinner2]?.flatten()))
+        Meal meal3 = new Meal(type: MealType.DINNER, items: Item.getAll([params?.meal?.Dinner3]?.flatten()))
+        Meal meal4 = new Meal(type: MealType.DINNER, items: Item.getAll([params?.meal?.Dinner4]?.flatten()))
+        Meal meal5 = new Meal(type: MealType.BREAKFAST, items: Item.getAll([params?.meal?.Breakfast]?.flatten()))
+        Meal meal6 = new Meal(type: MealType.LUNCH, items: Item.getAll([params?.meal?.Lunch]?.flatten()))
         quickFill.name = params.quickFillName
         quickFill.addToMealItems(meal1)
         quickFill.addToMealItems(meal2)
@@ -40,7 +45,11 @@ class QuickFillController {
         quickFill.addToMealItems(meal4)
         quickFill.addToMealItems(meal5)
         quickFill.addToMealItems(meal6)
-        quickFill.s()
+        try{
+            quickFill.s()
+        }catch(ex){
+
+        }
         flash.message = (params.id) ? "Record updated to database." : "Record saved to database."
         redirect(action: "quickFillAdmin")
     }
